requirements.txt
certifi==2019.3.9
chardet==3.0.4
docopt==0.6.2
idna==2.8
python-dotenv==0.10.2
requests==2.22.0
urllib3==1.25.2
